# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Mastori Api specification",
        "description": "",
        "contact": {
            "email": "giannaropoulou.foteini@gmail.com;tania.pets@gmail.com"
        },
        "version": "1.0.0"
    },
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "users"
                ],
                "description": "Returns all the end users",
                "operationId": "getEndUsers",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Access token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All end users response",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/endUser"
                            }
                        }
                    },
                    "400": {
                        "description": "token_not_provided",
                        "schema": {
                            "$ref": "#/definitions/errorModel"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "users"
                ],
                "description": "Adds a new end user in database",
                "operationId": "addEndUser",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/endUserPost"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Returns created end user object",
                        "schema": {
                            "$ref": "#/definitions/mastori"
                        }
                    },
                    "400": {
                        "description": "validation errors",
                        "schema": {
                            "$ref": "#/definitions/validationsErrorsModel"
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "put": {
                "tags": [
                    "users"
                ],
                "description": "Edits an end user",
                "operationId": "editenduser",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of end user to update",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Access token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/endUserPost"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "End user response",
                        "schema": {
                            "$ref": "#/definitions/endUser"
                        }
                    },
                    "401": {
                        "description": "token_not_provided/token_invalid",
                        "schema": {
                            "$ref": "#/definitions/errorModel"
                        }
                    },
                    "400": {
                        "description": "validation errors",
                        "schema": {
                            "$ref": "#/definitions/validationsErrorsModel"
                        }
                    }
                }
            }
        },
        "/users/{user_id}/favorites": {
            "get": {
                "tags": [
                    "users"
                ],
                "description": "Gets an end user's favorite mastoria",
                "operationId": "getUserFavorites",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Access token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID of end user",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns favorite mastoria",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/mastori"
                            }
                        }
                    },
                    "401": {
                        "description": "token_not_provided/token_invalid/Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errorModel"
                        }
                    }
                }
            }
        },
        "/favorites": {
            "post": {
                "tags": [
                    "favorites"
                ],
                "description": "Adds a mastori into favorite list",
                "operationId": "addFavorite",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Access token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mastoriId"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "validation errors",
                        "schema": {
                            "$ref": "#/definitions/validationsErrorsModel"
                        }
                    },
                    "401": {
                        "description": "token_not_provided/token_invalid/Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errorModel"
                        }
                    }
                }
            }
        },
        "/favorites/{mastori_id}": {
            "delete": {
                "tags": [
                    "favorites"
                ],
                "description": "Removes a mastori from favorites list",
                "operationId": "removefavorite",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "mastori_id",
                        "in": "path",
                        "description": "id of mastori to be removed",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Access token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "token_not_provided/token_invalid",
                        "schema": {
                            "$ref": "#/definitions/errorModel"
                        }
                    },
                    "500": {
                        "description": "Cannot be deleted.",
                        "schema": {
                            "$ref": "#/definitions/errorModel"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "address": {
            "required": [
                "lat",
                "lng",
                "address",
                "city",
                "country"
            ],
            "properties": {
                "id": {
                    "type": "integer",
                    "readOnly": true
                },
                "lat": {
                    "type": "number",
                    "format": "double"
                },
                "lng": {
                    "type": "number",
                    "format": "double"
                },
                "address": {
                    "type": "string"
                },
                "friendly_name": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string",
                    "format": "date-time",
                    "readOnly": true
                },
                "updated_at": {
                    "type": "string",
                    "format": "date-time",
                    "readOnly": true
                }
            }
        },
...
        "endUser": {
            "required": [
                "name",
                "phone"
            ],
            "properties": {
                "id": {
                    "type": "integer",
                    "readOnly": true
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "photo": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string",
                    "format": "date-time",
                    "readOnly": true
                },
                "updated_at": {
                    "type": "string",
                    "format": "date-time",
                    "readOnly": true
                },
                "favorites": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/mastori"
                    }
                }
            }
        },
        "endUserPost": {
            "required": [
                "name",
                "email",
                "phone",
                "password",
                "password_repeat",
                "addresses"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "password_repeat": {
                    "type": "string"
                },
                "photo": {
                    "type": "string"
                },
                "addresses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/address"
                    }
                }
            }
        },
...
        "errorModel": {
            "required": [
                "error"
            ],
            "properties": {
                "error": {
                    "type": "string",
                    "readOnly": true
                }
            }
        },
        "fieldValidationDescription": {
            "properties": {
                "error_description": {
                    "type": "string"
                }
            }
        },
        "validationerror": {
            "properties": {
                "fieldName": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fieldValidationDescription"
                    }
                }
            }
        },
        "validationsErrorsModel": {
            "required": [
                "errors"
            ],
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/validationerror"
                    }
                }
            }
        },
        "mastori": {
            "required": [
                "username",
                "email",
                "password",
                "last_name",
                "first_name",
                "phone",
                "pricelist"
            ],
            "properties": {
                "id": {
                    "type": "integer",
                    "readOnly": true
                },
                "username": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "format": "password"
                },
                "last_name": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "paratsoukli": {
                    "type": "string"
                },
                "facebook_id": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "pricelist": {
                    "type": "string"
                },
                "avg_response_time": {
                    "type": "integer"
                },
                "avg_rating": {
                    "type": "number",
                    "format": "float"
                },
                "active": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string",
                    "format": "date-time",
                    "readOnly": true
                },
                "updated_at": {
                    "type": "string",
                    "format": "date-time",
                    "readOnly": true
                },
                "addresses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/address"
                    }
                },
                "professions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/profession"
                    }
                }
            },
            "type": "object"
        },
...
        "mastoriId": {
            "properties": {
                "mastori_id": {
                    "type": "integer"
                }
            }
        },
        "profession": {
            "required": [
                "tag",
                "title"
            ],
            "properties": {
                "id": {
                    "type": "integer",
                    "readOnly": true
                },
                "tag": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string",
                    "format": "date-time",
                    "readOnly": true
                },
                "updated_at": {
                    "type": "string",
                    "format": "date-time",
                    "readOnly": true
                }
            }
        },
...
    }
}